pyproject.toml
[project]
name = "stock-intelligence"
version = "0.1.0"
description = "Macro-intelligent stock research agent"
requires-python = ">=3.11,<3.14"

dependencies = [
    "crewai>=0.28.0",
    "boto3>=1.34.0",
    "langchain-aws>=0.1.0",
    "pymongo>=4.6.0",
    "mcp>=1.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "schedule>=1.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pytz>=2024.1",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "tools", "db", "models", "mcp_servers"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
# lancedb 0.30.0 (crewai's transitive dep) has no macOS x86_64 wheel.
# Override to a newer release that ships macosx_11_0_x86_64 wheels.
override-dependencies = [
    "lancedb==0.25.3",       # 0.26+ dropped macOS x86_64 wheels
    "onnxruntime==1.23.2",   # 1.24+ dropped macOS x86_64 wheels
]
